% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shailaR.R
\docType{package}
\name{shailar}
\alias{shailar}
\title{shailaR: A package with a minimum example of package creation.}
\description{
This package is an exercise in package creation using
R studio. The package includes a sample function and
a sample dataset with their respective documentation.
}
\references{
\url{https://github.com/paezha/Reproducible-Research-Workflow}
}
\author{
Shaila Jamal, School of Geography and Earth Sciences, McMaster University \email{jamals16@mcmaster.ca}
}
